requirements.txt
Flask>=2.0
Flask-Session>=0.4
PyPDF2>=3.0
openai>=1.0
python-dotenv>=1.0
gunicorn>=20.0

streamlit==1.31.1
python-docx==1.0.1
PyMuPDF==1.23.26
Pillow==10.2.0
beautifulsoup4==4.12.3
streamlit-javascript
streamlit-pdf-viewer==0.0.4
streamlit-js-eval==0.1.5 